codex: rename injected message module
