+++ codex-rs/core/src/codex/timer_runtime.rs
@@ -31,13 +31,13 @@ use super::Event;
 use super::EventMsg;
 use super::INITIAL_SUBMIT_ID;
 use super::Session;
-use crate::messages::MessageInvocationContext;
-use crate::messages::MessageInvocationKind;
-use crate::messages::MessagePayload;
-use crate::messages::db_message_to_thread_message;
-use crate::messages::injected_message_event;
-use crate::messages::message_prompt_input_item;
-use crate::messages::validate_meta;
+use crate::injected_message::MessageInvocationContext;
+use crate::injected_message::MessageInvocationKind;
+use crate::injected_message::MessagePayload;
+use crate::injected_message::db_message_to_thread_message;
+use crate::injected_message::injected_message_event;
+use crate::injected_message::message_prompt_input_item;
+use crate::injected_message::validate_meta;
 use crate::pending_input::GeneratedMessageInput;
 use crate::pending_input::PendingInputItem;
 use crate::timers::ClaimedTimer;
