+++ codex-rs/core/src/injected_message.rs
